Hydrological models

Usage

model_linearReservoir(Q_In, Q_Out0 = 0, param_K = 1)

Arguments

Q_In

(vector) Inflow Boundary condition (Forcing) data

Q_Out0

(num) Innitial condition, Outflow in the first timepoint, DEFAULT in 0

param_K

(num) Parameter K of the linear reservoir method

Value

Outflow (vector)

Linear Reservoir

Linear Reservoir is a method and just assuming that the watershed behaves like a linear reservoir, where the outflow is proportional to the water storage within the reservoir.

\[Q_{out} = \frac{1}{K}S(t)\]

In addition to their relationship with output and storage, linear reservoir models also adhere to the continuity equation, often referred to as the water balance equation.

\[\frac{\mathrm{d}S(t)}{\mathrm{d}t} = Q_{in} - Q_{out}\]

By combining both equations, we obtain a differential equation (DGL).

\[Q_{in} = Q_{out} + K\frac{\mathrm{d}Q_{out}(t)}{\mathrm{d}t}\]

\[Q_{out}(t)=\int_{\tau=t0}^{t}Q_{in}(\tau)\frac{1}{K}e^{-\frac{t-\tau}{K}}\mathrm{d}\tau + Q_{out}(t_0)\frac{1}{K}e^{-\frac{t-t0}{K}}\]

Where:

\(Q_{in}\) is the inflow of the reservoir

\(Q_{out}\) is the outflow of the reservoir

\(S\) is the storage of the reservoir

\(K\) is the parameter that defines the relationship between $Q_out$ and $S$