Some tools for flow direction / path / net

This package provides some tools for handling flow direction, path, and network-related calculations.

Usage

get_inflow_cells(num_FlowDir)

read_UNF(fn_UNF)

read_UNF_scp(fn_UNF, ip_Host, str_Username, str_Password)

write_UNF(data_Export, fn_UNF)

Arguments

num_FlowDir

Integer. The cell number of the next cell. The cell number must range from 1 to the length of the cells. If the cell has no outflow, the number should be set to itself.

fn_UNF

string, name of the UNF file

ip_Host

string, the name of the remote host or its IP address

str_Username

string, the name of the user on the remote machine

str_Password

string, a password for accessing the local SSH key

data_Export

vector, data, that would exported

Value

A list of integers representing the cells that flow into the specified cell.

vector of numric or integer by read, UNF file by write

Functions

get_inflow_cells: Get all the cells which will flow into a specified cell (including the cell itself).

UNF Format

UNF0 / UNF8: numeric with 4 / 8 Bytes,

UNF1, UNF2 / UNF4: integer with 1, 2, / 4 Bytes

Read & Write

Read data from UNF file

Write data to UNF file

Examples

num_FD <- c(4,5,2,5,9,3,7,8,8)
lst_Cells <- get_inflow_cells(num_FD)